Added support for required properties
With the default options, an item has properties for a title, a description and a link, but these are optional. With this commit, the prevalue editor now let developers specify that a property should be required, hidden or optional (default). If a user in the backoffice has added a new item, but are missing one or more required properties, this is now also reflected in the UI.