Improved the logic around ImagePickerConfiguration
In previous releases, the ImagePickerConfiguration class has inherited from Umbraco's MediaPickerConfiguration class, as we'd then inherit relevant properties for the configuration. However as the "MediaPickerConfiguration" class also has other propeties - eg. "OnlyImages" and "DisableFolderSelect" - that are not relevant, the "ImagePickerConfiguration" class has now been updated to implement the relevant properties on it's own instead of inheriting them from "MediaPickerConfiguration".